fix: add 201 Created to status codes
The `Location` header is _also_ used by the `201 Created` status code.  But the currently logic filters out the information provided in `ctx.meta.$location` if I return `ctx.meta.$statusCode = 201`.  This should not happen.  The `Location` header is essential in a `201 Created` response.  I've changed to logic so that `201` is included in the status codes that passthrough `ctx.meta.$location` to the `Location` header.